backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
langchain>=0.0.200
chromadb>=0.3.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
pytest>=6.2.5
httpx>=0.24.0
aiofiles>=0.8.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0 